requirements.txt
black
matplotlib
py-cpuinfo